build:  buildsystem use conan to build Flexus
This commit introduce the usage of conan to build
Flexus. The commit also rewrite the CMake build to keep
only the essential, while also searching for the latest boost
library instead of looking for an absolute path. This remove
consequently Configuration.cmake. Test file and stats-manager
files were removed.

Flexus source code was modified to comply with the compiler
and new compile flags used.